correct bug mail in api + improve css
